+++ Runtime/Scripts/Authentication/CallbackHandler/ChangePasswordCallback.cs
@@ -3,7 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace UnityFirebaseREST
+namespace FirebaseRestClient
 {
     public class ChangePasswordCallback
     {