+++ Runtime/Scripts/Authentication/CallbackHandler/GetUserCallback.cs
@@ -3,14 +3,14 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace UnityFirebaseREST
+namespace FirebaseRestClient
 {
     public class GetUserCallback 
     {
-        internal Action<GetUserResponse> successCallback;
+        internal Action<FirebaseUser> successCallback;
         internal Action<Exception> exceptionCallback;
 
-        public GetUserCallback OnSuccess(Action<GetUserResponse> callback)
+        public GetUserCallback OnSuccess(Action<FirebaseUser> callback)
         {
             successCallback += callback;
             return this;
@@ -39,5 +39,23 @@ namespace UnityFirebaseREST
         public string lastRefreshAt;
         public bool disabled;
         public bool customAuth;
+
+        public FirebaseUser ToUser(FirebaseUser oldUser)
+        {
+            oldUser.localId = localId;
+            oldUser.email = email;
+            oldUser.isEmailVerified = emailVerified;
+            oldUser.displayName = displayName;
+            oldUser.photoUrl = photoUrl;
+
+            oldUser.validSince = validSince;
+            oldUser.lastLoginAt = lastLoginAt;
+            oldUser.createdAt = createdAt;
+            oldUser.lastRefreshAt = lastRefreshAt ;
+            oldUser.disabled = disabled;
+            oldUser.customAuth = customAuth;
+
+            return oldUser;
+        }
     }
 }