+++ Runtime/Scripts/Authentication/CallbackHandler/GoogleIdTokenCallback.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace UnityFirebaseREST
+namespace FirebaseRestClient
 {
     public class GoogleIdTokenCallback
     {