+++ Runtime/Scripts/Authentication/CallbackHandler/SignUpCallback.cs
@@ -3,14 +3,14 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace UnityFirebaseREST
+namespace FirebaseRestClient
 {
     public class SignUpCallback 
     {
-        internal Action<SignUpResponse> successCallback;
+        internal Action<FirebaseUser> successCallback;
         internal Action<Exception> exceptionCallback;
 
-        public SignUpCallback OnSuccess(Action<SignUpResponse> callback)
+        public SignUpCallback OnSuccess(Action<FirebaseUser> callback)
         {
             successCallback += callback;
             return this;
@@ -33,5 +33,16 @@ namespace UnityFirebaseREST
         public string localId;
 
         public SignUpResponse() { }
+
+        public FirebaseUser ToUser()
+        {
+            var tempUser = new FirebaseUser();
+            tempUser.localId = localId;
+            tempUser.email = email;
+            tempUser.accessToken = idToken;
+            tempUser.refreshToken = refreshToken;
+
+            return tempUser;
+        }
     }
 }