+++ Runtime/Scripts/Authentication/CallbackHandler/UpdateUserCallback.cs
@@ -3,14 +3,14 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace UnityFirebaseREST
+namespace FirebaseRestClient
 {
     public class UpdateUserCallback
     {
-        internal Action<UpdateProfileResponse> successCallback;
+        internal Action<FirebaseUser> successCallback;
         internal Action<Exception> exceptionCallback;
 
-        public UpdateUserCallback OnSuccess(Action<UpdateProfileResponse> callback)
+        public UpdateUserCallback OnSuccess(Action<FirebaseUser> callback)
         {
             successCallback += callback;
             return this;
@@ -24,6 +24,7 @@ namespace UnityFirebaseREST
     }
     public class UpdateProfileResponse
     {
+        public string localId;
         public string email;
         public string displayName;
         public string photoUrl;
@@ -33,6 +34,18 @@ namespace UnityFirebaseREST
         public int expiresIn;
 
         public UpdateProfileResponse() { }
+
+        public FirebaseUser ToUser(FirebaseUser oldUser)
+        {
+            oldUser.localId = localId;
+            oldUser.email = email;
+            oldUser.displayName = displayName;
+            oldUser.photoUrl = photoUrl;
+            oldUser.accessToken = idToken;
+            oldUser.refreshToken = refreshToken;
+
+            return oldUser;
+        }
     }
 }