+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## [0.5.2] - 01-04-2021
+  
+User related Functions returns user object, namespace name changed to FirebaseRestClient, StateChanged event and minor changes.
+
+### Added
+- User object callback
+- StateChanged event
+- SignOut
+- Responses to User Object (Internal Only)
+- Key-Value Pair Writing to Realtime Database
+- Efficient way to assign Filters
+
+### Changed
+- Namespace name changed
+
+## [0.5.1] - 29-03-2021
+  
+Fixed Props names.
+
+### Added
+- CurrentUser and OnStateChanged props name change
+
+
 ## [0.5.0] - 29-03-2021
   
 Initially added Realtime Database and Firebase Authentication (Experimental) support. 
@@ -13,3 +36,4 @@ Initially added Realtime Database and Firebase Authentication (Experimental) sup
 - User Profile Email Change
 - User Profile Password Change
 - User Profile Link with Email or Providers
+